Feat/extension

+++ packages/app/src/store/daidai.ts
@@ -1,20 +1,21 @@
-import { supabaseClient, User } from "@supabase/auth-helpers-nextjs"
+import { User } from "@supabase/auth-helpers-nextjs"
 import create from "zustand"
 import { immer } from "zustand/middleware/immer"
 
-import { daidaisQuery } from "../types/api"
 import { ANONYMOUS_DAIDAIS, isAnonymousDaidai } from "../utils/anonymous"
+import { daidaisQuery, supabaseClient } from "../utils/supabaseClient"
 import DaidaiObject from "./DaidaiObject"
 import logger from "./plugins/logger"
 import { selectTags } from "./selector"
 
 export type DaidaiState = {
+  user: User | null
   data: DaidaiObject[]
   activeTags: string[]
 
   initDatda: (user: User | null) => Promise<void>
   reset: (daidaiObjects: DaidaiObject[]) => void
-  add: (userId: string, ...daidaiObjects: DaidaiObject[]) => Promise<void>
+  add: (...daidaiObjects: DaidaiObject[]) => Promise<void>
   update: (index: number, daidaiObject: DaidaiObject) => Promise<void>
   remove: (index: number) => Promise<void>
   toggleTag: (tag: string) => void
@@ -23,13 +24,14 @@ export type DaidaiState = {
 
 const useDaiDaiStore = create<DaidaiState>()(
   logger(
-    // persist(
     immer((set, get) => ({
+      user: null,
       data: [],
       activeTags: [],
 
       initDatda: async (user) => {
         if (user) {
+          // online mode
           const result = await daidaisQuery().select("id, url, c_html")
           if (result.error) throw result.error
           const daidaisJSON = [
@@ -37,15 +39,28 @@ const useDaiDaiStore = create<DaidaiState>()(
             ...result.data,
           ]
           set({
+            user,
             data: daidaisJSON.map(
               (item) => new DaidaiObject({ url: item.url, contentHTML: item.c_html, id: item.id })
             ),
           })
         } else {
-          set({
-            data: ANONYMOUS_DAIDAIS.map(
+          // offline mode
+          console.time("init with localStorage")
+          const localStr = localStorage.getItem(LOCAL_KEY)
+          let _data
+          if (localStr) {
+            const result = JSON.parse(localStr)
+            _data = result.state.data.map((item: any) => DaidaiObject.hydrate(item)).filter(Boolean)
+          } else {
+            _data = ANONYMOUS_DAIDAIS.map(
               (item) => new DaidaiObject({ url: item.url, contentHTML: item.c_html, id: item.id })
-            ),
+            )
+          }
+          console.timeEnd("init with localStorage")
+          set({
+            user,
+            data: _data,
           })
         }
       },
@@ -54,18 +69,32 @@ const useDaiDaiStore = create<DaidaiState>()(
           data: daidaiObjects,
         })
       },
-      add: async (userId, ...daidaiObjects) => {
-        const { error } = await daidaisQuery().insert(
-          daidaiObjects.map((item) => ({
-            url: item.url,
-            c_html: item.contentHTML,
-            user_id: userId,
-          }))
-        )
-        if (error) throw error
-        set((state) => {
-          state.data.push(...daidaiObjects)
-        })
+      add: async (...daidaiObjects) => {
+        const state = get()
+        if (state.user) {
+          const { error, data } = await daidaisQuery()
+            .insert(
+              daidaiObjects.map((item) => ({
+                url: item.url,
+                c_html: item.contentHTML,
+                user_id: state.user!.id,
+              }))
+            )
+            .select("id, url, c_html")
+          if (error) throw error
+          // use database id
+          set((state) => {
+            state.data.push(
+              ...data.map(
+                (item) => new DaidaiObject({ url: item.url, contentHTML: item.c_html, id: item.id })
+              )
+            )
+          })
+        } else {
+          set((state) => {
+            state.data.push(...daidaiObjects)
+          })
+        }
       },
       update: async (index, daidaiObject) => {
         const state = get()
@@ -74,14 +103,16 @@ const useDaiDaiStore = create<DaidaiState>()(
         if (isAnonymousDaidai(preDaidaiObject.id)) {
           return
         }
-        const { error } = await daidaisQuery()
-          .update({
-            url: daidaiObject.url,
-            c_html: daidaiObject.contentHTML,
-          })
-          .eq("id", preDaidaiObject.id)
-        console.error("e", error)
-        if (error) throw error
+        if (state.user) {
+          const { error } = await daidaisQuery()
+            .update({
+              url: daidaiObject.url,
+              c_html: daidaiObject.contentHTML,
+            })
+            .eq("id", preDaidaiObject.id)
+          console.error("e", error)
+          if (error) throw error
+        }
         set((state) => {
           if (index in state.data) {
             state.data[index] = daidaiObject
@@ -92,16 +123,18 @@ const useDaiDaiStore = create<DaidaiState>()(
         const state = get()
         const daidaiObject = state.data[index]
         if (!daidaiObject) return
-        if (isAnonymousDaidai(daidaiObject.id)) {
-          const { error } = await supabaseClient.auth.update({
-            data: {
-              [daidaiObject.id]: true,
-            },
-          })
-          if (error) throw error
-        } else {
-          const { error } = await daidaisQuery().delete().eq("id", daidaiObject.id)
-          if (error) throw error
+        if (state.user) {
+          if (isAnonymousDaidai(daidaiObject.id)) {
+            const { error } = await supabaseClient.auth.updateUser({
+              data: {
+                [daidaiObject.id]: true,
+              },
+            })
+            if (error) throw error
+          } else {
+            const { error } = await daidaisQuery().delete().eq("id", daidaiObject.id)
+            if (error) throw error
+          }
         }
         set((state) => {
           state.data.splice(index, 1)
@@ -131,29 +164,23 @@ const useDaiDaiStore = create<DaidaiState>()(
         })
       },
     }))
-    // {
-    //   name: "__daidai",
-    //   serialize: (localState) => {
-    //     return JSON.stringify({
-    //       ...localState,
-    //       state: {
-    //         ...omit(localState.state, "activeTags"),
-    //         data: localState.state.data.map((item) => item.dehydrate()),
-    //       },
-    //     })
-    //   },
-    //   deserialize: (str) => {
-    //     console.time("init with localStorage")
-    //     const result = JSON.parse(str)
-    //     result.state.data = result.state.data
-    //       .map((item: any) => DaidaiObject.hydrate(item))
-    //       .filter(Boolean)
-    //     console.timeEnd("init with localStorage")
-    //     return result
-    //   },
-    // }
-    // )
   )
 )
+
+const LOCAL_KEY = "__DAIDAI__"
+
+useDaiDaiStore.subscribe((state, prevState) => {
+  if (!state.user && state.data !== prevState.data) {
+    localStorage.setItem(
+      LOCAL_KEY,
+      JSON.stringify({
+        state: {
+          // ...omit(localState.state, "activeTags"),
+          data: state.data.map((item) => item.dehydrate()),
+        },
+      })
+    )
+  }
+})
 
 export default useDaiDaiStore
